[6주차] 컴포넌트 분리
